requirements.txt
fastapi
uvicorn
facenet-pytorch
pillow
requests
numpy
torch
